get_statistics_subset should return well-formatted null results
The existing code for handling lists of values that don't meet minimum
length causes problems when the `percentiles` key is used; e.g., rather
than the expected

    [{percentile, [{0.5, 0.0}, ... ]}]

the user is presented with

    [{{percentile, [0.5, ... ]}, 0.0}]

which doesn't match the formatting for other subset key. This patch
special-cases the `percentile` key to return the expected result.